refactor using worker

+++ src/JobProcessor.php
@@ -2,41 +2,30 @@
 
 namespace winwin\jobQueue;
 
-use RuntimeException;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\EventDispatcher\GenericEvent;
 
 class JobProcessor implements JobProcessorInterface
 {
     /**
-     * @var JobQueueInterface
+     * @var EventDispatcherInterface
      */
-    private $jobQueue;
-    
+    private $eventDispatcher;
+
     /**
      * @var string
      */
     private $pidfile;
 
     /**
-     * @var EventDispatcherInterface
+     * @var LockHandler
      */
-    private $eventDispatcher;
-
-    /**
-     * @var JobFactoryInterface
-     */
-    private $jobFactory;
+    private $lock;
 
     /**
      * @var int
      */
-    private $maxRequests;
-
-    /**
-     * @var int
-     */
-    private $workers;
+    private $lastHeartbeatTime;
 
     /**
      * @var bool
@@ -53,27 +42,39 @@ class JobProcessor implements JobProcessorInterface
      */
     private $workerPids = [];
 
-    public function __construct(JobQueueInterface $jobQueue, JobFactoryInterface $jobFactory, EventDispatcherInterface $eventDispatcher, $pidfile, $workers = 1, $maxRequests = 100)
+    /**
+     * @var WorkerInterface[]
+     */
+    private $workers;
+
+    public function __construct(EventDispatcherInterface $eventDispatcher, $pidfile, $heartbeatInterval = 60)
     {
         if (strpos(strtolower(PHP_OS), 'win') === 0) {
-            throw new RuntimeException("This application not support windows");
+            throw new \RuntimeException("This application not support windows");
         }
 
         // 检查扩展
         if (!extension_loaded('pcntl')) {
-            throw new RuntimeException("Please install pcntl extension");
+            throw new \RuntimeException("Please install pcntl extension");
         }
 
         if (!extension_loaded('posix')) {
-            throw new RuntimeException("Please install posix extension");
+            throw new \RuntimeException("Please install posix extension");
         }
 
-        $this->jobQueue = $jobQueue;
         $this->eventDispatcher = $eventDispatcher;
-        $this->jobFactory = $jobFactory;
         $this->pidfile = $pidfile;
-        $this->workers = $workers;
-        $this->maxRequests = $maxRequests;
+        $this->lock = new LockHandler(md5($pidfile), $heartbeatInterval);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function addWorker(WorkerInterface $worker)
+    {
+        $this->workers[] = $worker;
+
+        return $this;
     }
 
     /**
@@ -83,15 +84,25 @@ class JobProcessor implements JobProcessorInterface
     {
         declare(ticks=1);
 
+        if (empty($this->workers)) {
+            throw new \RuntimeException("No workers added.");
+        }
+
         $this->checkProcess();
         $this->eventDispatcher->dispatch(Events::PROCESSOR_START, new GenericEvent($this));
         $this->installSignal();
         try {
             while (!$this->stopped) {
+                if (!$this->heartbeat()) {
+                    throw new \RuntimeException("Heartbeat failed");
+                }
                 $this->startWorkers();
                 $pid = pcntl_wait($status, WNOHANG);
                 if ($pid) {
-                    unset($this->workerPids[$pid]);
+                    $key = array_search($pid, $this->workerPids);
+                    if ($key) {
+                        unset($this->workerPids[$key]);
+                    }
                 } else {
                     sleep(1);
                 }
@@ -99,6 +110,7 @@ class JobProcessor implements JobProcessorInterface
         } catch (\Exception $e) {
             $this->stopWorkers();
         }
+        $this->lock->release();
         unlink($this->pidfile);
     }
 
@@ -124,8 +136,16 @@ class JobProcessor implements JobProcessorInterface
         if ($pid) {
             return posix_kill($pid, SIGUSR1);
         } else {
-            throw new RuntimeException("Job processor is not running");
+            throw new \RuntimeException("Job processor is not running");
         }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function isAlive()
+    {
+        return $this->lock->isAlive();
     }
 
     protected function getPid()
@@ -135,31 +155,49 @@ class JobProcessor implements JobProcessorInterface
         }
     }
 
+    /**
+     * check whether the processor already started
+     */
     protected function checkProcess()
     {
-        $pid = $this->getPid();
-        if ($pid > 0 && posix_kill($pid, 0)) {
-            throw new RuntimeException("Job processor is running, pid=$pid");
+        if ($this->lock->lock(false)) {
+            $this->lastHeartbeatTime = time();
+            $pid = $this->getPid();
+            if ($pid > 0 && posix_kill($pid, 0)) {
+                throw new \RuntimeException("Job processor is running, pid=$pid");
+            }
+            $dir = dirname($this->pidfile);
+            if (!is_dir($dir) && !mkdir($dir, 0777, true)) {
+                throw new \RuntimeException("Cannot create pid file directory '$dir'");
+            }
+            file_put_contents($this->pidfile, getmypid());
+        } else {
+            throw new \RuntimeException("Lock exists, is the processor running?");
         }
-        $dir = dirname($this->pidfile);
-        if (!is_dir($dir) && !mkdir($dir, 0777, true)) {
-            throw new RuntimeException("Cannot create pid file directory '$dir'");
+    }
+
+    protected function heartbeat()
+    {
+        if (time() - $this->lastHeartbeatTime > $this->lock->getHeartbeatInterval() - 5) {
+            return $this->lock->heartbeat();
         }
-        file_put_contents($this->pidfile, getmypid());
+        return true;
     }
 
     protected function startWorkers()
     {
-        while (count($this->workerPids) < $this->workers) {
-            $pid = pcntl_fork();
-            if ($pid == -1) {
-                throw new RuntimeException('Cannot fork queue worker');
-            } elseif ($pid) {
-                $this->workerPids[$pid] = true;
-            } else {
-                $this->reinstallSignal();
-                $this->startWorker();
-                exit;
+        foreach ($this->workers as $i => $worker) {
+            if (!isset($this->workerPids[$i])) {
+                $pid = pcntl_fork();
+                if ($pid == -1) {
+                    throw new \RuntimeException('Cannot fork queue worker');
+                } elseif ($pid) {
+                    $this->workerPids[$i] = $pid;
+                } else {
+                    $this->reinstallSignal();
+                    $this->startWorker($worker);
+                    exit;
+                }
             }
         }
     }
@@ -215,83 +253,22 @@ class JobProcessor implements JobProcessorInterface
 
     protected function stopWorkers()
     {
-        foreach ($this->workerPids as $workerPid => $ignore) {
+        foreach ($this->workerPids as $i => $workerPid) {
             if (posix_kill($workerPid, SIGINT)) {
                 pcntl_waitpid($workerPid, $status);
-                unset($this->workerPids[$workerPid]);
+                unset($this->workerPids[$i]);
             }
         }
     }
 
-    protected function startWorker()
+    protected function startWorker(WorkerInterface $worker)
     {
-        $this->eventDispatcher->dispatch(Events::WORKER_START, new GenericEvent($this));
-        $processedJobs = 0;
-        while (!$this->workerStopped && $processedJobs < $this->maxRequests) {
-            $job = $this->jobQueue->reserve(1);
-            if (!$job) {
-                continue;
-            }
-            $event = new GenericEvent($this);
-            $event['job'] = $job;
-            try {
-                $data = json_decode($job->getData(), true);
-                if (!isset($data['job'], $data['payload'])) {
-                    throw new \UnexpectedValueException('invalid job');
-                }
-                if (!class_exists($data['job'])) {
-                    throw new \UnexpectedValueException("job {$data['job']} does not exist");
-                }
-                $consumer = $this->jobFactory->create($data['job']);
-                if (!$consumer instanceof JobInterface) {
-                    throw new \UnexpectedValueException("job {$data['job']} does not implement ".JobInterface::class);
-                }
-                $event['consumer'] = $consumer;
-                $event['payload'] = $data['payload'];
-                $this->eventDispatcher->dispatch(Events::BEFORE_PROCESS_JOB, $event);
-                $consumer->process($data['payload']);
-                $this->jobQueue->delete($job);
-                $this->eventDispatcher->dispatch(Events::AFTER_PROCESS_JOB, $event);
-                ++$processedJobs;
-            } catch (\Pheanstalk\Exception $e) {
-                $event['error'] = $e;
-                $this->eventDispatcher->dispatch(Events::JOB_FAILED, $event);
-                sleep(100);
-            } catch (\Exception $e) {
-                $event['error'] = $e;
-                $this->eventDispatcher->dispatch(Events::JOB_FAILED, $event);
-                $this->jobQueue->bury($job);
-            } catch (\Error $e) {
-                $event['error'] = $e;
-                $this->eventDispatcher->dispatch(Events::JOB_FAILED, $event);
-                $this->jobQueue->bury($job);
-            }
+        $this->eventDispatcher->dispatch(Events::WORKER_START, new GenericEvent($worker));
+        $worker->start();
+        while (!$this->workerStopped && $worker->shouldRun()) {
+            $worker->work();
         }
-        $this->eventDispatcher->dispatch(Events::WORKER_STOP, new GenericEvent($this));
-    }
-
-    public function getMaxRequests()
-    {
-        return $this->maxRequests;
-    }
-
-    public function setMaxRequests($maxRequests)
-    {
-        $this->maxRequests = $maxRequests;
-        return $this;
-    }
-
-    public function getWorkers()
-    {
-        return $this->workers;
-    }
-
-    public function setWorkers($workers)
-    {
-        $this->workers = (int) $workers;
-        if ($this->workers < 1) {
-            $this->workers = 1;
-        }
-        return $this;
+        $worker->stop();
+        $this->eventDispatcher->dispatch(Events::WORKER_STOP, new GenericEvent($worker));
     }
 }
